add impl_arithmetic macros
